g1_admin_bot/requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.5
asyncio>=3.4.3
python-dotenv>=1.0.0 